Move service classes to services package
