+++ code/triangle/View/src/main/java/ru/unn/agile/Triangle/view/Triangle.java
@@ -2,13 +2,13 @@ package ru.unn.agile.triangle.view;
 
 import javafx.beans.property.ReadOnlyListProperty;
 import javafx.collections.ListChangeListener;
-import javafx.collections.ObservableList;
 import javafx.event.ActionEvent;
 import javafx.event.EventHandler;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
 import javafx.scene.control.TextField;
+import ru.unn.agile.triangle.view.services.ViewModelProvider;
 import ru.unn.agile.triangle.viewmodel.LoggerRecordViewModel;
 import ru.unn.agile.triangle.viewmodel.ViewModel;
