+++ code/triangle/View/src/main/java/ru/unn/agile/triangle/view/services/ViewModelProvider.java
@@ -1,8 +1,9 @@
-package ru.unn.agile.triangle.view;
+package ru.unn.agile.triangle.view.services;
 
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.ReadOnlyObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
+import ru.unn.agile.triangle.view.services.LoggerService;
 import ru.unn.agile.triangle.viewmodel.ViewModel;
 
 public class ViewModelProvider {
